feat: absorb hub/store/PWA implementations; reconcile status docs
This tree had been living inside the bogusyogi/claude workspace with no .git of
its own, so it was disconnected from this remote. Reconnected, with the work done
while disconnected brought back here.

Implementations absorbed (they existed only on branches of that workspace repo,
which have since been deleted after their content was preserved):
- full axum hub: auth/http/router/state/ws + auth/delivery/http/reconnect suites,
  replacing the 18-line healthz/readyz stub
- roundtable-store: real implementation over the 11-table schema, replacing the
  20-line open()+migrations stub
- roundtable-protocol: expanded types, PROTOCOL_VERSION widened u8 -> u16
- packages/web: the real PWA — api.ts, offline.ts IndexedDB queue, and the
  Composer/Login/MessageList/RoomList/RoomView/SeatPanel components, replacing the
  42-line static demo

cargo test --workspace: 62 passed, 0 failed (was 30).

CORRECTION: an earlier pass through these docs called the CHANGELOG push history
(3455587, 2b3c0c2, 2c3140f, 6e5ac7a, 210047f) fabricated because those objects
could not be found. That was wrong — all five are ancestors of main HERE. The
check had been run from the disconnected copy, against the wrong repository. The
claim is reinstated as accurate in README, SUMMARY, CHANGELOG, and STATUS.

KNOWN GAP, not hidden: roundtable-node and roundtable-hub speak different wire
framings. The protocol's WsEnvelope flattens its event; node's client expects a
nested `payload` under a `type` tag. Node keeps a private Envelope<T> in hub.rs,
documented in place, so its 24 tests and fake-hub.mjs stay honest about the
framing node actually speaks. Swapping in WsEnvelope without porting the fixtures
and adding a node<->hub test would give a green build that fails on the wire.

Also:
- STATUS.md added as the single authoritative status document
- three-ring synthesis architecture doc added
- cargo workspace members listed explicitly; a globbed crates/* picked up tooling
  scratch dirs and failed the whole workspace
- .agent/reconcile.json -> .agent/stale.json (the former never existed)

+++ tools/roundtable/crates/roundtable-node/src/lib.rs
@@ -24,7 +24,7 @@ pub enum NodeError {
     #[error("hub rejected hello: {0}")]
     HubRejected(String),
     #[error("protocol version mismatch: hub={hub} node={node}")]
-    ProtocolVersion { hub: u8, node: u8 },
+    ProtocolVersion { hub: u16, node: u16 },
     #[error("invalid frame: {0}")]
     InvalidFrame(String),
     #[error("provider error: {0}")]
